IndexedDB/idbobjectstore_get5.htm

<!DOCTYPE html>
<html>
<head>
    <meta content="text/html; charset=utf-8" http-equiv="content-type" />
    <title> IDBObjectStore.get() - returns the record with the first key in the range </title>
    <script type="text/javascript" src="support.js"></script>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script type="text/javascript">
        var objectStoreName = "objectstore";
        var db = null;
        
        var t = async_test();
        
        function RunTest() {
            var rqOpen = window.indexedDB.open(databaseName, databaseVersion);
            rqOpen.onupgradeneeded = t.step_func(function(event) {
                db = event.target.result;
                db.onerror = t.step_func(assert_database_error);
                var objStore = db.createObjectStore(objectStoreName, {keyPath:"key"});
                for(var i = 0; i < 10; i++) {
                    objStore.add({key:i, property:"data"});
                }
            });
            rqOpen.onerror = t.step_func(assert_open_request_error);
            rqOpen.onsuccess = t.step_func(GetRecordFromObjectStore);
        }
        
        function GetRecordFromObjectStore() {
            var txn = db.transaction(objectStoreName, "readonly");
            var objStore = txn.objectStore(objectStoreName);
            var rqGet = objStore.get(IDBKeyRange.bound(3,6));
            rqGet.onsuccess = t.step_func(function(event) {
                assert_equals(event.target.result.key, 3, 'event.target.result.key');
                t.done();
            });
        }
        
        add_completion_callback(function() { if(db) db.close(); });
        
        setup(function() {
            var rqDelete = window.indexedDB.deleteDatabase(databaseName);
            rqDelete.onsuccess = t.step_func( RunTest );
        });
    </script>
</head>
<body>
    <div id="log"></div>
</body>
</html>
